Renamed  .carousel
Renamed the section .carousel to .carousel-box as there was another class in the code from bootstrap with the same name.
Gave the carousel a box-shadow and
darkened the shade of the hero-image by adding the linear-gradient property to enable the letters of the navigation links pop (hitherto, they didn't have much constrast with the background).