Changes to HTML and CSS
Header changed from a background colour to an image.
More images added on to HTML to create a grid of photos.
Hidden nav put in a div with hamburger menu icon.
Footer styled to improve appearance.
Margin changes in CSS to make use of white space and some media queries added