I added css grid to position my elements
I added CSS grid to position the elements inside the body. I added all my images in the main and all I have left to do is add some css grid to the header to position the elements in the right place and add some info in the footer element.